+++ src/main/kotlin/net/serverwars/sunsetPlugin/domain/gameservertype/services/GameServerTypeDataAccess.kt
@@ -1,33 +1,20 @@
 package net.serverwars.sunsetPlugin.domain.gameservertype.services
 
-import io.ktor.client.call.*
 import io.ktor.client.request.*
-import io.ktor.http.*
-import net.serverwars.sunsetPlugin.Main
 import net.serverwars.sunsetPlugin.config.Config
 import net.serverwars.sunsetPlugin.domain.gameservertype.models.gameservertypelist.GameServerTypeList
 import net.serverwars.sunsetPlugin.domain.gameservertype.models.gameservertypelist.GameServerTypeListDto
 import net.serverwars.sunsetPlugin.domain.gameservertype.services.mappers.GameServerTypeListMapper
 import net.serverwars.sunsetPlugin.util.rest.HttpClient
-import net.serverwars.sunsetPlugin.util.rest.exceptions.ApiException
+import net.serverwars.sunsetPlugin.util.rest.exceptions.parse
 
 object GameServerTypeDataAccess {
 
     suspend fun getAvailableGameServerTypes(): GameServerTypeList {
         val url = "${Config.getApiBaseUrl()}/gameServerTypes"
-        return runCatching {
-            val response = HttpClient.instance.get(url)
-
-            if (!response.status.isSuccess()) {
-                error("${response.status} Response body: ${response.body() ?: ""}")
-            }
-
-            val gameServerTypeListDto = response.body<GameServerTypeListDto>()
-            GameServerTypeListMapper.fromDto(gameServerTypeListDto)
-        }.getOrElse { error ->
-            Main.inst.logger.severe("[API EXCEPTION] Could not get available game server types: ${error.message ?: "Unknown error"}")
-            throw ApiException()
-        }
+        val response = HttpClient.instance.get(url)
+        val dto = response.parse<GameServerTypeListDto>("Could not get available game server types")
+        return GameServerTypeListMapper.fromDto(dto)
     }
 
 }