+++ src/main/kotlin/net/serverwars/sunsetPlugin/domain/gameservertype/services/GameServerTypeService.kt
@@ -1,15 +1,15 @@
 package net.serverwars.sunsetPlugin.domain.gameservertype.services
 
 import net.serverwars.sunsetPlugin.domain.gameservertype.models.gameservertype.GameServerType
-import net.serverwars.sunsetPlugin.util.runAsync
+import net.serverwars.sunsetPlugin.util.rest.exceptions.callApi
 
 object GameServerTypeService {
 
     val availableGameServerTypes = mutableListOf<GameServerType>()
 
     fun refreshAvailableGameServerTypes() {
-        runAsync {
-            this.availableGameServerTypes.clear()
+        this.availableGameServerTypes.clear()
+        callApi {
             this.availableGameServerTypes.addAll(GameServerTypeDataAccess.getAvailableGameServerTypes().gameServerTypes)
         }
     }