+++ src/main/kotlin/net/serverwars/sunsetPlugin/domain/match/services/MatchDataAccess.kt
@@ -1,10 +1,9 @@
 package net.serverwars.sunsetPlugin.domain.match.services
 
-import io.ktor.client.call.*
 import io.ktor.client.plugins.sse.*
 import io.ktor.client.request.*
 import io.ktor.http.*
-import io.ktor.utils.io.CancellationException
+import io.ktor.utils.io.*
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.cancel
@@ -22,7 +21,7 @@ import net.serverwars.sunsetPlugin.domain.match.services.mappers.MatchListMapper
 import net.serverwars.sunsetPlugin.domain.match.services.mappers.MatchStatusEventMapper
 import net.serverwars.sunsetPlugin.domain.server.services.ServerService
 import net.serverwars.sunsetPlugin.util.rest.HttpClient
-import net.serverwars.sunsetPlugin.util.rest.exceptions.ApiException
+import net.serverwars.sunsetPlugin.util.rest.exceptions.parse
 import net.serverwars.sunsetPlugin.util.rest.parseSSEDto
 import java.util.*
 import kotlin.time.Duration.Companion.minutes
@@ -31,24 +30,14 @@ object MatchDataAccess {
 
     suspend fun getActiveMatchList(): MatchList {
         val url = "${Config.getApiBaseUrl()}/matches"
-        return runCatching {
-            val response = HttpClient.instance.get(url) {
-                parameters {
-                    parameter("filter_server_uuid", ServerService.getServerUuid().toString())
-                    parameter("filter_in_session", true)
-                }
+        val response = HttpClient.instance.get(url) {
+            parameters {
+                parameter("filter_server_uuid", ServerService.getServerUuid().toString())
+                parameter("filter_in_session", true)
             }
-
-            if (!response.status.isSuccess()) {
-                error("${response.status} Response body: ${response.body() ?: ""}")
-            }
-
-            val matchListDto = response.body<MatchListDto>()
-            MatchListMapper.fromDto(matchListDto)
-        }.getOrElse { error ->
-            Main.inst.logger.severe("[API EXCEPTION] Could not get active matches: ${error.message ?: "Unknown error"}")
-            throw ApiException()
         }
+        val dto = response.parse<MatchListDto>("Could not get active matches")
+        return MatchListMapper.fromDto(dto)
     }
 
     @OptIn(FlowPreview::class)