+++ src/main/kotlin/net/serverwars/sunsetPlugin/domain/queue/services/QueueService.kt
@@ -10,8 +10,6 @@ import net.serverwars.sunsetPlugin.domain.match.services.MatchService
 import net.serverwars.sunsetPlugin.domain.menu.models.menu.LobbyMenu
 import net.serverwars.sunsetPlugin.domain.menu.models.menuitem.QueueEnterMenuItem
 import net.serverwars.sunsetPlugin.domain.menu.models.menuitem.QueueLeaveMenuItem
-import net.serverwars.sunsetPlugin.domain.queue.exceptions.QueueEnterException
-import net.serverwars.sunsetPlugin.domain.queue.exceptions.QueueLeaveException
 import net.serverwars.sunsetPlugin.domain.queue.models.queueentrystatus.QueueEntryStatus
 import net.serverwars.sunsetPlugin.domain.queue.models.queueentrystatus.QueueEntryStatusType
 import net.serverwars.sunsetPlugin.domain.queue.services.mappers.QueueEntryCreateMapper
@@ -20,7 +18,8 @@ import net.serverwars.sunsetPlugin.util.now
 import net.serverwars.sunsetPlugin.util.playEnterQueueSound
 import net.serverwars.sunsetPlugin.util.playLeaveQueueSound
 import net.serverwars.sunsetPlugin.util.rest.exceptions.ApiException
-import net.serverwars.sunsetPlugin.util.runAsync
+import net.serverwars.sunsetPlugin.util.rest.exceptions.ErrorCodeMessage
+import net.serverwars.sunsetPlugin.util.rest.exceptions.callApi
 import net.serverwars.sunsetPlugin.util.runSync
 import org.bukkit.Bukkit
 import org.bukkit.Material
@@ -39,93 +38,110 @@ object QueueService {
     fun getTimeInQueue(): Long = this.queueEnterTimestamp?.let { now() - it } ?: 0
 
     fun enterQueue(audience: Audience): CompletableFuture<Unit> {
-        if (this.cooldownTaskId != null) return CompletableFuture.completedFuture(Unit)
+        val default = CompletableFuture.completedFuture(Unit)
 
-        audience.sendTranslatedMessage("command.queue.entering")
-        return runAsync {
-            try {
-                sendEnterQueue()
-                addCooldown(this.QUEUE_COOLDOWN_AFTER_ENTER, QueueLeaveMenuItem.material)
-            } catch (error: QueueEnterException) {
-                audience.sendTranslatedMessage(error.key, *error.args)
-            }
+        // Check cooldown
+        if (this.cooldownTaskId != null) {
+            return default
         }
-    }
 
-    fun leaveQueue(audience: Audience = Audience.empty()): CompletableFuture<Unit> {
-        if (this.cooldownTaskId != null) return CompletableFuture.completedFuture(Unit)
-
-        audience.sendTranslatedMessage("command.queue.leaving")
-        return runAsync {
-            try {
-                sendLeaveQueue()
-                addCooldown(this.QUEUE_COOLDOWN_AFTER_LEAVE, QueueEnterMenuItem.material)
-            } catch(error: QueueLeaveException) {
-                audience.sendTranslatedMessage(error.key, *error.args)
-            }
-        }
-    }
-
-    private suspend fun sendEnterQueue() {
+        // Check not in queue yet
         if (this.queueUuid != null) {
-            throw QueueEnterException("command.queue.enter.error.already_in_queue")
+            audience.sendTranslatedMessage("command.queue.enter.error.already_in_queue")
+            return default
         }
 
-        val lobby = LobbyService.getLobbyCopy() ?: throw QueueEnterException("command.queue.enter.error.no_lobby")
+        // Check lobby exists
+        val lobby = LobbyService.getLobbyCopy()
+        if (lobby == null) {
+            audience.sendTranslatedMessage("command.queue.enter.error.no_lobby")
+            return default
+        }
 
+        // Check minimum amount of players in lobby
         if (lobby.getParticipantAmount() < Lobby.MIN_LOBBY_SIZE) {
-            throw QueueEnterException(
+            audience.sendTranslatedMessage(
                 "command.queue.enter.error.not_enough_players",
                 lobby.getParticipantAmount(),
                 Lobby.MIN_LOBBY_SIZE
             )
+            return default
         }
 
-        // Make sure server is not in match yet
-        try {
+        audience.sendTranslatedMessage("command.queue.entering")
+        return callApi(
+            audience = audience,
+            errorCodeMap = mapOf(
+                "QUEUE_ENTRY_VALIDATION_ALREADY_IN_MATCH" to ErrorCodeMessage("command.queue.enter.error.already_in_match"),
+            )
+        ) {
+            // Make sure server is not in match yet
             val isInMatch = MatchService.checkInMatch()
             if (isInMatch) {
-                throw QueueEnterException("command.queue.enter.error.already_in_match")
+                audience.sendTranslatedMessage(
+                    "command.queue.enter.error.already_in_match",
+                    lobby.getParticipantAmount(),
+                    Lobby.MIN_LOBBY_SIZE
+                )
+                return@callApi
             }
-        } catch (_: ApiException) {
-            throw QueueEnterException("command.queue.enter.error.api_exception")
-        }
 
-        // Enter queue
-        val queueEntryCreate = QueueEntryCreateMapper.fromLobby(lobby)
-        try {
+            // Enter queue
+            val queueEntryCreate = QueueEntryCreateMapper.fromLobby(lobby)
             val queueEntryCreateResponse = QueueDataAccess.enterQueue(queueEntryCreate)
             this.queueUuid = queueEntryCreateResponse.queueEntryUuid
             QueueTimerService.startTimer(lobby, queueEntryCreateResponse.queueEntryUuid)
-        } catch (_: ApiException) {
-            throw QueueEnterException("command.queue.enter.error.api_exception")
-        }
-        lobby.sendMessage("command.queue.enter.success.notify_lobby")
-        playEnterQueueSound(lobby)
-        LobbyStatusNotifierService.stopShowingLobbyStatus()
 
-        this.queueEnterTimestamp = now()
-        Main.inst.logger.info("[QUEUE] Entering queue, searching for opponents...")
+            // Notify lobby
+            lobby.sendMessage("command.queue.enter.success.notify_lobby")
+            playEnterQueueSound(lobby)
+            LobbyStatusNotifierService.stopShowingLobbyStatus()
+
+            this.queueEnterTimestamp = now()
+            Main.inst.logger.info("[QUEUE] Entering queue, searching for opponents...")
+            addCooldown(this.QUEUE_COOLDOWN_AFTER_ENTER, QueueLeaveMenuItem.material)
+        }
     }
 
-    private suspend fun sendLeaveQueue() {
-        val queueUuidCopy = this.queueUuid ?: throw QueueLeaveException("command.queue.leave.error.not_in_queue")
+    fun leaveQueue(audience: Audience = Audience.empty()): CompletableFuture<Unit> {
+        val default = CompletableFuture.completedFuture(Unit)
 
+        // Check cooldown
+        if (this.cooldownTaskId != null) {
+            return default
+        }
+
+        // Check in queue
+        val queueUuidCopy = this.queueUuid
+        if (queueUuidCopy == null) {
+            audience.sendTranslatedMessage("command.queue.leave.error.not_in_queue")
+            return default
+        }
+
+        // Remove queue uuid
         this.queueUuid = null
         QueueTimerService.stopTimer(queueUuidCopy)
 
-        try {
+        audience.sendTranslatedMessage("command.queue.leaving")
+        return callApi(
+            audience = audience,
+            errorCodeMap = mapOf(
+                "QUEUE_ENTRY_NOT_IN_QUEUE" to ErrorCodeMessage("command.queue.leave.error.not_in_queue"),
+            )
+        ) {
+            // Leave queue
             QueueDataAccess.leaveQueue(queueUuidCopy)
 
+            // Notify lobby
             LobbyService.getLobbyCopy()?.let {
                 it.sendMessage("command.queue.leave.success.notify_lobby")
                 playLeaveQueueSound(it)
                 LobbyStatusNotifierService.startShowingLobbyStatus()
             }
+
             Main.inst.logger.info("[QUEUE] Left queue, time in queue: ${getTimeInQueue()}ms")
             this.queueEnterTimestamp = null
-        } catch (_: ApiException) {
-            throw QueueLeaveException("command.queue.leave.error.api_exception")
+            addCooldown(this.QUEUE_COOLDOWN_AFTER_LEAVE, QueueEnterMenuItem.material)
         }
     }
 
@@ -144,12 +160,12 @@ object QueueService {
             val matchUuid = queueEntryStatus.matchUuid!!
             MatchService.setMatch(matchUuid)
             MatchDataAccess.listenToMatchStatusEvents(matchUuid = matchUuid)
-        }
-        else if (queueEntryStatus.status == QueueEntryStatusType.LEFT_QUEUE && this.queueUuid != null) {
+        } else if (queueEntryStatus.status == QueueEntryStatusType.LEFT_QUEUE && this.queueUuid != null) {
             LobbyService.getLobbyCopy()?.sendMessage("command.queue.leave.error.forced_to_leave_queue")
             try {
-                sendLeaveQueue()
-            } catch (_: QueueLeaveException) { }
+                leaveQueue()
+            } catch (_: ApiException) {
+            }
         }
     }