+++ src/main/kotlin/net/serverwars/sunsetPlugin/domain/server/services/ServerService.kt
@@ -3,7 +3,7 @@ package net.serverwars.sunsetPlugin.domain.server.services
 import net.serverwars.sunsetPlugin.config.Config
 import net.serverwars.sunsetPlugin.domain.server.exceptions.GetServerException
 import net.serverwars.sunsetPlugin.domain.server.models.Server
-import net.serverwars.sunsetPlugin.util.runAsync
+import net.serverwars.sunsetPlugin.util.rest.exceptions.callApi
 import java.util.UUID
 
 object ServerService {
@@ -11,16 +11,16 @@ object ServerService {
     private var server: Server? = null
 
     fun reloadServer() {
-        runAsync {
+        callApi {
             val serverSecret = Config.getServerSecret()
             this.server = ServerDataAccess.getServerFromSecret(serverSecret)
         }
     }
 
     fun getServerUuid(): UUID = this.server?.serverUuid
-        ?: throw GetServerException("Cached server is null. Is the secret correctly configured?")
+        ?: throw GetServerException("Cached server is null. Is the secret correctly configured in the plugin's config?")
 
     fun getServerSlug(): String = this.server?.slug
-        ?: throw GetServerException("Cached server is null. Is the secret correctly configured?")
+        ?: throw GetServerException("Cached server is null. Is the secret correctly configured in the plugin's config?")
 
 }