+++ src/main/kotlin/net/serverwars/sunsetPlugin/util/rest/exceptions/ApiErrorResponse.kt
@@ -1,0 +1,9 @@
+package net.serverwars.sunsetPlugin.util.rest.exceptions
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ApiErrorResponse(
+    val errorCode: String,
+    val message: String
+)