+++ src/main/kotlin/net/serverwars/sunsetPlugin/util/rest/exceptions/ApiException.kt
@@ -1,3 +1,9 @@
 package net.serverwars.sunsetPlugin.util.rest.exceptions
 
-open class ApiException : Exception()
+import io.ktor.http.HttpStatusCode
+
+class ApiException(
+    val status: HttpStatusCode,
+    val errorCode: String,
+    override val message: String,
+) : Exception(message)